+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v1.0.2
 * Fix: selling relics no longer buys the card behind the click
 * Fix: work around basemod hack that made selling potions seem impossible
+* Most event relics can no longer be sold.
 
 ### v1.0.1
 * Fix relic description